+++ src/index.ts
@@ -2,7 +2,7 @@ export type {Task, TaskContext} from './task/task.js';
 export type {Gen, GenContext} from './gen/gen.js';
 
 // export all of the main config helpers and types
-export {loadGroConfig, toConfig} from './config/config.js';
+export {loadConfig, toConfig} from './config/config.js';
 export type {
 	GroConfig,
 	GroConfigPartial,
@@ -11,11 +11,7 @@ export type {
 	GroConfigCreatorOptions,
 } from './config/config.js';
 // also export the build config stuff
-export {
-	normalizeBuildConfigs,
-	validateBuildConfigs,
-	isPrimaryBuildConfig,
-} from './build/buildConfig.js';
+export {normalizeBuildConfigs, validateBuildConfigs} from './build/buildConfig.js';
 export type {BuildConfig, BuildName, BuildConfigPartial} from './build/buildConfig.js';
 
 // these seem useful and generic enough to export to users